refactor: rewrite MetaSprite Tileset Compiler

+++ src/models/metasprite/compiler/combinesmalltilesets.h
@@ -13,7 +13,7 @@
 namespace UnTech {
 namespace MetaSprite {
 namespace Compiler {
-typedef std::unordered_map<unsigned, std::array<unsigned, 4>> SmallTileMap_t;
+typedef std::unordered_map<unsigned, std::array<uint16_t, 4>> SmallTileMap_t;
 
 SmallTileMap_t combineSmallTilesets(const TileGraph_t& smallTileGraph);
 }
